requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
cryptography==42.0.5
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==5.0.3
django-appconf==1.0.6
django-cryptography==1.1
django-filter==24.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.0.11
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
locust==2.24.0
MarkupSafe==2.1.5
msgpack==1.0.8
oauthlib==3.2.2
packaging==24.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg==3.1.18
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
pyzmq==25.1.2
redis==5.0.3
requests==2.31.0
requests-oauthlib==1.4.0
roundrobin==0.0.4
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
whitenoise==6.6.0
zope.event==5.0
zope.interface==6.2